Reworked initial test code, added level movement and test character
Added the level movement, but couldn't get it working in C#, so did it in the default Godot scripting language. Build kept crashing when attempting to run it with both .cs and .gd files, so had to rework the initial mouse movement test code to be in the Godot script as well. Added a test character to start working on the player controls and direct gameplay, but got stuck for the moment.