Added player character controls through testController.gd.
Added movement for the player character- now the player model (represented by the old man with axe) will go to an upper area, a middle area, or a lower area based on what section of the screen the mouse is in once a button on the keyboard is pressed. Character will stay in that position until a button is pressed while the mouse is in a different area. By default the boundaries of the areas are y-values of 200 and 400. Will need to code in a way to differentiate between running and covering. Cory assisted with brainstorming in voice chat while I wrote out this code.

Co-Authored-By: dep <9596170+EnilaDep@users.noreply.github.com>