Adjusted character controls and added the beginning of damage detection.
Modified the character controls script to be a bit more streamlined as well as adjusting for any changes in viewport size. Added a sunlight area and "sunlight" group as well as preliminary code to the character controller to interpret damage from entering the sunlight while outside of the "cover" state.