Added level background, finished obstacles, added damage, added level music, and staged level.
Went through and added the background to the level. Figured out how to get collision and damage calculation working. Staged the level with the obstacles, may need adjusted for difficulty. Added spaces to account for losing the game and winning the game in the PC code, but will need to add it once the menu's been added, which is hinging on tubar sending the assets some point today. Added the music to the level as well, no SFX yet, not sure if we wanted to do that. Wrote some notes down in a .txt file going over how to get the systems I set up today working based off my own research.